Fix build interrupt breaking subsequent builds
Much of the time spent compiling is during the postgresql source tarball
download. If the build were interrupted during this process, subsequent
builds would fail since the postgresql source was only partially
downloaded, but the Makefile thought it was.

This patch corrects that behavior by downloading the source to a tmp
folder first. The extracted source is moved to its destination only
after it is completely available.

Resolves #25.